Use manual test since iotest.TestReader is only available in Go 1.16
